memorialCalender/requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.6
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
idna==2.10
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
pylint==2.6.0
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
wrapt==1.12.1
djangorestframework-jwt==1.11.0